requirements.txt
aniso8601==9.0.1
arabic-reshaper==2.1.3
asttokens==2.0.8
backcall==0.2.0
beautifulsoup4==4.11.1
Brotli==1.0.9
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cssselect2==0.7.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.0.0
Flask==2.2.2
fonttools==4.37.3
fpdf==1.7.2
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.4
ipykernel==6.15.3
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
numpy==1.23.3
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydyf==0.3.0
Pygments==2.13.0
pyparsing==3.0.9
pyphen==0.13.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.2.1
pyzmq==24.0.0
qrcode==7.3.1
reportlab==3.6.11
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.5.0
tinycss2==1.1.1
tornado==6.2
traitlets==5.4.0
urllib3==1.26.12
wcwidth==0.2.5
weasyprint==56.1
webencodings==0.5.1
Werkzeug==2.2.2
zopfli==0.2.1
